% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_normalization.R
\name{countMatrix_norm}
\alias{countMatrix_norm}
\title{Normalize a count matrix}
\usage{
countMatrix_norm(count.matrix, DESeq = FALSE)
}
\arguments{
\item{count.matrix}{raw count matrix (genes are rows, columns are samples)}

\item{DESeq}{logical; if set to FALSE (default), edgeR normFactors will used for normalization; otherwise, DEseq sizeFactors will be used}
}
\value{
normalized count matrix
}
\description{
This function takes a count matrix, calculates either DESeq sizeFactors or EdgeR normFactors and normalizes counts using limma::voom
}
\examples{
countMatrix_norm(expr)
}
