% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_normalization.R
\name{counts2TPM}
\alias{counts2TPM}
\title{Transform count matrix to tmp matrix}
\usage{
counts2TPM(count.matrix, gene.length.ob)
}
\arguments{
\item{count.matrix}{raw count matrix (genes are rows, columns are samples); rownames are esemble ID}

\item{gene.length.ob}{object get from gene_length function; a dataframe of geneID(esmbleID) and gene length; rownames are esemble ID}
}
\value{
a tpm matrix (genes are rows, columns are samples); rownames are esemble ID
}
\description{
Transform count matrix to tmp matrix
}
\examples{
counts2TPM(count.matrix, gene.length.ob)
}
