% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseq_DEG_analysis.R
\name{DEGsig}
\alias{DEGsig}
\title{Adding differatiation column}
\usage{
DEGsig(df, logFC, pV, method = "Deseq2")
}
\arguments{
\item{df}{DEG data frame analysed with deseq2 limma or edgeR}

\item{logFC}{log2(fold change)}

\item{pV}{pvalue}

\item{method}{Deseq2 limma or edgeR (default Deseq2)}
}
\value{
DEG dataframe adding a column named Dif
}
\description{
Adding differatiation column
}
\examples{
DEGsig(df, logFC =1, pV = 0.05, method = 'Deseq2')
}
