% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseq_DEG_analysis.R
\name{filter_count_matrix}
\alias{filter_count_matrix}
\title{Filtering a count matrix to remove low read genes}
\usage{
filter_count_matrix(matrix, threshold = 0.1, minCpm = 0.25)
}
\arguments{
\item{threshold}{minimum persentage of samples in which CPM threshold has to be reached; defaults to ten percent (0.1)}

\item{minCpm}{the minimum counts per million (CPM) that has to be reached; defaults to ten percent (0.25)}

\item{mmatrix}{raw count matrix (genes are rows, columns are samples)}
}
\value{
filtered count matrix
}
\description{
This function takes a count matrix and removes genes that don't reach a minimum namer of reads in a certain number of samples.
}
\examples{
filter_count_matrix(matrix, threshold = 0.1, minCpm=0.25)
}
