% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_normalization.R
\name{gene_length}
\alias{gene_length}
\title{get a dataframe of gene length from a GTF file}
\usage{
gene_length(gtf.file.path)
}
\arguments{
\item{gtf.file.path}{the file path of target GTF file}
}
\value{
a dataframe of geneID(esmbleID) and gene length
}
\description{
get a dataframe of gene length from a GTF file
}
\examples{
gene_length(gtf.file.path = '../../!公共资源/gencode.v36.annotation.gtf.gz')
}
