% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO_KEGG.R
\name{GO_barplot_catergroy}
\alias{GO_barplot_catergroy}
\title{barplot (ordered by count/ratio) and group by category of ontology (BP, MF, CC)}
\usage{
GO_barplot_catergroy(
  GO.res,
  top = 10,
  y_value = "count",
  decreasing.Y = TRUE,
  color.text = F,
  lab.size = 4,
  y.limit,
  pal = (ggsci::pal_lancet(palette = c("lanonc"), alpha = 0.6))(9),
  bar.width = 0.6,
  title = NULL,
  xlab = NULL,
  ylab = "Gene counts",
  base.size = 15,
  legend.size = 10,
  legend.position = "top",
  cord.flip = F
)
}
\arguments{
\item{GO.res}{GO result object  generated from my_GO() function including (BP, MF, CC)}

\item{top}{number of top pathways picked, default 10}

\item{y_value}{ordered by count or ratio, default "count", if need ratio, use  y_value = 'ratio'}

\item{decreasing.Y}{ordered by decreasing,default T}

\item{color.text}{if using colored labels, (go terms generated by geom_text, better give y.limit value) / if not show go terms on axis text}

\item{lab.size}{text size of colored labels, default 4}

\item{y.limit}{a vector with 2 numbers as the ymin and ymax: e.g y.limit = c(-30, 40), if not using default value}

\item{pal}{color palette represent for up or down regulation; down as the first color; up as the second color ;default in 'lanonc' pal}

\item{bar.width}{bar width ,default 0.6}

\item{title}{plot title ,default NULL}

\item{xlab}{plot title ,default NULL}

\item{ylab}{plot title ,default count / ratio}

\item{base.size}{base text size, default 15}

\item{legend.size}{base legend size, default 10}

\item{legend.position}{legend positon, default top}

\item{cord.flip}{if flip the plot default F}
}
\value{
a sorted ggplot bar plot grouped by GO terms
}
\description{
barplot (ordered by count/ratio) and group by category of ontology (BP, MF, CC)
}
\examples{
GO_barplot_catergroy(GO.res = GOResults_list[[1]], color.text = T, y.limit = c(-30, 45))

}
