% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO_KEGG.R
\name{GO_barplot_catergroy_facet}
\alias{GO_barplot_catergroy_facet}
\title{barplot (ordered by count/ratio) and group by category of ontology with facets of BP CC MF}
\usage{
GO_barplot_catergroy_facet(
  GO.res,
  top = 20,
  y_value = "count",
  y.decreasing = T,
  lab.position = "above_bar",
  lab.size = 4,
  lab.colour = "black",
  pal = c("#25A17C", "#D96622", "#6C73B2"),
  title = NULL,
  ylab = "Gene counts",
  y.lab.size = 14,
  y.title.size = 18,
  title.size = 20
)
}
\arguments{
\item{GO.res}{GO result object  generated from my_GO() function including (all,BP, MF, CC), the all (list[\link{1}]) is used}

\item{top}{number of top pathways picked, default 20}

\item{y_value}{ordered by count or ratio, default "count", if need ratio, use  y_value = 'ratio'}

\item{y.decreasing}{if use decreasing order, default T}

\item{lab.position}{the position of the labs of GO terms; default above_bar: above the bar; or on_bar : upon the bar}

\item{lab.size}{the text size the labs, default 4}

\item{lab.colour}{the text colour the labs, default black}

\item{pal}{color palette represent for BP, MF, CC  GO category}

\item{title}{title of the plot , default NULL}

\item{ylab}{title of the y axis, default Gene counts}

\item{y.lab.size}{text size of the y labs, default 14}

\item{y.title.size}{text size of the y title, default 18}

\item{title.size}{text size of the title, default 20}
}
\value{
a ggplot object o facet GO plot
}
\description{
barplot (ordered by count/ratio) and group by category of ontology with facets of BP CC MF
}
\examples{
GO_barplot_catergroy_facet(GO.res = GO.up, top = 20, y_value = 'count', y.decreasing = T, lab.position = 'on_bar', lab.colour = 'white')
}
