% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_GSEA.R
\name{GSEA.plot_simple}
\alias{GSEA.plot_simple}
\title{GSEA plot a simple version; given one pathway name interested}
\usage{
GSEA.plot_simple(pathway, GSEA.res, base_size = 4)
}
\arguments{
\item{pathway}{pathway name , as shown in the gsea analysis}

\item{GSEA.res}{gsea result object, not a data frame!}

\item{base_size}{font size}
}
\value{
a simple version of gsea plot of a single pathway
}
\description{
GSEA plot a simple version; given one pathway name interested
}
\examples{
GSEA.plot_simple(pathway, GSEA.res)
}
