% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_GSEA.R
\name{GSEA_barplot1}
\alias{GSEA_barplot1}
\title{GSEA presented as barplot ; given top pathways arranged by pvalue}
\usage{
GSEA_barplot1(
  GESA.res,
  top = 30,
  pal = (ggsci::pal_lancet(palette = c("lanonc"), alpha = 0.6))(9),
  title = "GSEA",
  legend.position = "right"
)
}
\arguments{
\item{GESA.res}{gsea result object, not a data frame!}

\item{top}{number of top pathways picked, default 30}

\item{pal}{color palette represent for up or down regulation; down as the first color; up as the second color ;default in 'lanonc' pal}

\item{title}{title of the plot , default GSEA}

\item{legend.position}{position of the legend, default right}
}
\value{
bar plot of GSEA result
}
\description{
GSEA presented as barplot ; given top pathways arranged by pvalue
}
\examples{
my_GSEA_barplot1(GSEAres)
}
