% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_GSEA.R
\name{GSEA_barplot2}
\alias{GSEA_barplot2}
\title{GSEA presented as barplot ; given top pathways arranged by pvalue, pathway name beside the bar}
\usage{
GSEA_barplot2(
  GESA.res,
  top = 30,
  pal = (ggsci::pal_lancet(palette = c("lanonc"), alpha = 0.6))(9),
  text.size = 4,
  xlab.size = 15,
  legendtext.size = 12
)
}
\arguments{
\item{GESA.res}{gsea result object, not a data frame!}

\item{top}{number of top pathways picked, default 30}

\item{pal}{color palette represent for up or down regulation; down as the first color; up as the second color ;default in 'lanonc' pal}

\item{text.size}{text size of pathway names,default 4}

\item{xlab.size}{text size of xlab,default 15}

\item{legendtext.size}{text size of legend,default 12}
}
\value{
bar plot of GSEA result
}
\description{
GSEA presented as barplot ; given top pathways arranged by pvalue, pathway name beside the bar
}
\examples{
my_GSEA_barplot2(GSEAres)
}
