% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_GSEA.R
\name{GSEA_chickletBar}
\alias{GSEA_chickletBar}
\title{GSEA presented as chicklet-barplot ; given top pathways arranged by pvalue, pathway name beside the bar}
\usage{
GSEA_chickletBar(
  GESA.res,
  top = 30,
  pal = (ggsci::pal_simpsons())(2),
  text.size = 3,
  xlab.size = 15,
  legendtext.size = 12
)
}
\arguments{
\item{GESA.res}{gsea result object, not a data frame!}

\item{top}{number of top pathways picked, default 30}

\item{pal}{color palette represent for up or down regulation; down as the first color; up as the second color ;default pal_simpsons}

\item{text.size}{text size of pathway names,default 3}

\item{xlab.size}{text size of xlab,default 15}

\item{legendtext.size}{text size of legend,default 12}
}
\value{
chicklet-bar plot of GSEA result
}
\description{
GSEA presented as chicklet-barplot ; given top pathways arranged by pvalue, pathway name beside the bar
}
\examples{
GSEA_chickletBar(GESA.res, top=30)
}
