% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSet_scoring.R
\name{gsScore}
\alias{gsScore}
\title{Calculate score across genes and samples using PCA for a matrix with genes in a defined geneset}
\usage{
gsScore(gm, summarizationFunction = "PCA")
}
\arguments{
\item{gm}{normalized count matrix; rows are all genes in the signature that shall be summarized into one score; columns are samples}

\item{summarizationFunction}{character vector defining whether signature scores shall be based on principal component 1 ("PCA", default) or z-scores (other value)}
}
\value{
numeric vector of length ncol(gm); a score summarizing the rows of gm
}
\description{
Calculate score across genes and samples using PCA for a matrix with genes in a defined geneset
}
