% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_GSVA.R
\name{GSVA_barplot1}
\alias{GSVA_barplot1}
\title{limma result of DEG from GSVA scoring matrix  presented as barplot}
\usage{
GSVA_barplot1(
  gsvaScore.limma.DEG,
  logFC = 0.2,
  pV = 0.05,
  addline = F,
  axis_need_color = T,
  palettes = c("#008020", "#808080", "#08519C")
)
}
\arguments{
\item{gsvaScore.limma.DEG}{limma result of DEG generated from GSVA scoring matrix}

\item{logFC}{log2(fold change) , deflaut 0.2}

\item{pV}{pvalue deflaut 0.05}

\item{addline}{if adding a helping line}

\item{axis_need_color}{if adding color on the axis lab of pathway name; showing up or down}

\item{palettes}{color palettes as down, stable, up}
}
\value{
bar plot of DEG using limma from a GSVA score matrix
}
\description{
limma result of DEG from GSVA scoring matrix  presented as barplot
}
\examples{
GSVA_barplot1(gsvaScore.limma.DEG)
}
