% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_GSVA.R
\name{GSVA_barplot2}
\alias{GSVA_barplot2}
\title{limma result of DEG from GSVA scoring matrix  presented as barplot, pathway name beside the bar}
\usage{
GSVA_barplot2(
  gsvaScore.limma.DEG,
  top = 20,
  pal = (ggsci::pal_lancet(palette = c("lanonc"), alpha = 0.6))(9),
  text.size = 4,
  xlab.size = 15,
  legendtext.size = 12
)
}
\arguments{
\item{gsvaScore.limma.DEG}{limma result of DEG generated from GSVA scoring matrix}

\item{top}{number of top pathways picked, default 20}

\item{pal}{color palette represent for up or down regulation; down as the first color; up as the second color ;default in 'lanonc' pal}

\item{text.size}{text size of pathway names,default 4}

\item{xlab.size}{text size of xlab,default 15}

\item{legendtext.size}{text size of legend,default 12}
}
\value{
bar plot of GSVA result
}
\description{
limma result of DEG from GSVA scoring matrix  presented as barplot, pathway name beside the bar
}
\examples{
GSVA_barplot2(gsvaScore.limma.DEG)
}
