% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_GSVA.R
\name{GSVA_volcano}
\alias{GSVA_volcano}
\title{limma result of DEG from GSVA scoring matrix  presented as vacano plot}
\usage{
GSVA_volcano(
  gsvaScore.limma.DEG,
  logFC = 0.2,
  pV = 0.05,
  pal = c("#DC143C", "#808080", "#00008B"),
  title = "GSVA",
  x.name = "log2 (Fold Change)",
  y.name = "-log10 (P-value)",
  labeled.pathway
)
}
\arguments{
\item{gsvaScore.limma.DEG}{limma result of DEG generated from GSVA scoring matrix}

\item{logFC}{log2(fold change) , deflaut 0.2}

\item{pV}{pvalue deflaut 0.05}

\item{labeled.pathway}{pathway name to be labeled}
}
\value{
volcano plot of DEG using limma from a GSVA score matrix
}
\description{
limma result of DEG from GSVA scoring matrix  presented as vacano plot
}
\examples{
GSVA_volcano(gsvaScore.limma.DEG)
}
