% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_normalization.R
\name{limitRange}
\alias{limitRange}
\title{Limit max or min value in the matrix}
\usage{
limitRange(m, max.z = 3)
}
\arguments{
\item{m}{normalized or scalized matrix of exprSet}

\item{max.z}{max z score}
}
\value{
limit max value to max.z p; min value to -max.z
}
\description{
Limit max or min value in the matrix
}
\examples{
limitRange(matirx, max.z = 3)
}
