% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_plotting.R
\name{my_barplot_long}
\alias{my_barplot_long}
\title{barplot for long dataframe}
\usage{
my_barplot_long(
  df,
  x,
  y,
  group,
  title = "",
  x.name = "",
  y.name = "",
  x_level,
  group_level,
  text_adding = T,
  text_size = 5,
  pal = (ggsci::pal_jco(alpha = 0.7))(9),
  plot_theme
)
}
\arguments{
\item{df}{dataframe of long format}

\item{x}{column name of x axis data}

\item{y}{column name of y axis data}

\item{group}{column name of grouping; presenting bar color}

\item{title}{title, default blank}

\item{x.name}{x lab name, default blank}

\item{y.name}{y lab name, default blank}

\item{x_level}{x lab levels, if x presenting as given order}

\item{group_level}{group lab levels, if group presenting as given order (show in legend figure)}

\item{text_adding}{if need adding number of the colnum; default true}

\item{text_size}{size of the text adding}

\item{pal}{color, default jco; if given notice the number of factors}

\item{plot_theme}{theme function returned object (e.g., my_themeFULL(), or theme_bw())}
}
\value{
ggplot2 barplot
}
\description{
barplot for long dataframe
}
\examples{
my_barplot_long(df, x, y, group)
}
