% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_plotting.R
\name{my_boxplot_long}
\alias{my_boxplot_long}
\title{boxplot for long dataframe}
\usage{
my_boxplot_long(
  df,
  x,
  y,
  group,
  title = "",
  x.name = "",
  y.name = "",
  x_level,
  group_level,
  pair_lab,
  jitter.point = T,
  jitter.size = 1,
  pal = (ggsci::pal_jco(alpha = 0.7))(9),
  plot_theme
)
}
\arguments{
\item{df}{dataframe of long format}

\item{x}{column name of x axis data}

\item{y}{column name of y axis data}

\item{group}{column name of grouping; presenting bar color}

\item{title}{title, default blank}

\item{x.name}{x lab name, default blank}

\item{y.name}{y lab name, default blank}

\item{x_level}{x lab levels, if x presenting as given order}

\item{group_level}{group lab levels, if group presenting as given order (show in legend figure)}

\item{pair_lab}{column name of the paried information}

\item{jitter.point}{if jitter points added; default true.}

\item{jitter.size}{size of jitter point}

\item{pal}{color, default jco; if given notice the number of factors}

\item{plot_theme}{theme function returned object (e.g., my_themeFULL(), or theme_bw())}
}
\value{
ggplot2 box plot object
}
\description{
boxplot for long dataframe
}
\examples{
my_boxplot_long(df, x, y, group)
}
