% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO_KEGG.R
\name{my_GO}
\alias{my_GO}
\title{GO analysis for human gene symbols; human only!}
\usage{
my_GO(DEG.genes, background.genes)
}
\arguments{
\item{DEG.genes}{gene symbols of DEGs}

\item{background.genes}{background gene symbols ,default not provide}
}
\value{
a list consists of GO result objects, inculuding all, BP, MF, CC
}
\description{
GO analysis for human gene symbols; human only!
}
\examples{
my_GO(deg.sig)
}
