% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_GSEA.R
\name{my_GSEA_analysis}
\alias{my_GSEA_analysis}
\title{GSEA analysis for human gene symbols}
\usage{
my_GSEA_analysis(DEG, fc_column, gene_column, GMTfile)
}
\arguments{
\item{DEG}{DEG dataframe from RNAseq results such as Deseq2}

\item{fc_column}{column name of fold change in the DEG}

\item{gene_column}{column name of genesymbol in the DEG}

\item{GMTfile}{gmt object of picked geneset gmt file}
}
\value{
GSEA object
}
\description{
GSEA analysis for human gene symbols
}
\examples{
my_GSEA_analysis(DEG, 'log2FoldChange', 'geneID', GMT.hallmark)
}
