% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_GSEA.R
\name{my_GSEAPlot}
\alias{my_GSEAPlot}
\title{Optimized GSEA plot version; given one pathway name interested}
\usage{
my_GSEAPlot(path, GSEA_Object, leading_genes)
}
\arguments{
\item{path}{pathway name , as shown in the gsea analysis}

\item{GSEA_Object}{gsea result object, not a data frame!}

\item{leading_genes}{leading edge gene symbols}
}
\value{
a simple version of gsea plot of a single pathway
}
\description{
Optimized GSEA plot version; given one pathway name interested
}
\examples{
my_GSEAPlot(path = "HALLMARK_EPITHELIAL_MESENCHYMAL_TRANSITION", GSEA_Object = gseaRes)
}
