% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_GSVA.R
\name{my_GSVA_analysis}
\alias{my_GSVA_analysis}
\title{GSVA analysis}
\usage{
my_GSVA_analysis(exprSet, GMTfile)
}
\arguments{
\item{exprSet}{A normlized expression matrix (e.g. TPM or FPKM, not raw count); genes are rows, columns are samples}

\item{GMTfile}{gmt object of picked geneset gmt file}
}
\value{
GSVA transformed geneset scored matrix (pathways are rows, columns are samples)
}
\description{
GSVA analysis
}
\examples{
my_GSVA_analysis(exprSet, GMTfile)
}
