% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseq_DEG_analysis.R
\name{my_hclust}
\alias{my_hclust}
\title{clusting visulization for RNAseq grouping}
\usage{
my_hclust(matrix)
}
\arguments{
\item{matrix}{raw count matrix (genes are rows, columns are samples)}
}
\value{
hclust plot
}
\description{
clusting visulization for RNAseq grouping
}
\examples{
my_hclust(exprSet)
}
