% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO_KEGG.R
\name{my_kegg}
\alias{my_kegg}
\title{KEGG analysis for human gene symbols; human only!}
\usage{
my_kegg(DEG.genes)
}
\arguments{
\item{DEG.genes}{gene symbols of DEGs}
}
\value{
kegg result object
}
\description{
KEGG analysis for human gene symbols; human only!
}
\examples{
my_kegg(deg)
}
