% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_plotting.R
\name{my_matchstick_plot}
\alias{my_matchstick_plot}
\title{matchstick_plot for go,kegg,gsea... analysis}
\usage{
my_matchstick_plot(
  df,
  top = 20,
  y.decreasing = T,
  logPV10 = T,
  x = "Description",
  y = "count",
  size = "count",
  fill_col = "pvalue",
  fill_col_pal = c("#BDD8E8", "#134B6C"),
  stick_col = "#FAAC90",
  stick_width = 2,
  title = NULL,
  xlab = NULL,
  ylab = "Gene counts",
  x.lab.size = 14,
  x.lab.angle = 0,
  y.lab.size = 14,
  y.lab.angle = 0,
  x.title.size = 18,
  y.title.size = 18,
  title.size = 20,
  col.limits = NULL,
  col.breaks,
  col.labels,
  size.limits = NULL,
  size.breaks,
  size.labels
)
}
\arguments{
\item{df}{dataframe of go, kegg,... results}

\item{top}{top term picked for plotting , arranged by count and pvalue , default top 20}

\item{y.decreasing}{if arranged the sticked with derceasing order, default T}

\item{logPV10}{if use the -log10pvalue for filling colors, default T}

\item{x}{column name of x axis data, (the plot is coordfliped) default Description}

\item{y}{column name of y axis data, (the plot is coordfliped) default count}

\item{size}{column name of point size, default count}

\item{fill_col}{column name of filling color, default pvalue}

\item{fill_col_pal}{color palette of the bulbe color; a vector consists of two numbers, the fisrt for low , the second for high;}

\item{stick_col}{color of the stick}

\item{stick_width}{width of the stick, default 2}

\item{title}{title of the plot, default NULL}

\item{xlab}{x lab of the plot, default NULL}

\item{ylab}{y lab of the plot, default 'Gene Counts'}

\item{x.lab.size}{text size of the x labs, default 14}

\item{x.lab.angle}{text angle of the x labs, default 0}

\item{y.lab.size}{text size of the y labs, default 14}

\item{y.lab.angle}{text angle of the y labs, default 0}

\item{x.title.size}{text size of the x lab title, default 18}

\item{y.title.size}{text size of the y lab title, default 18}

\item{title.size}{text size of the plot title, default 20}

\item{col.limits}{limit number of the color bar, default NULL, if given, a vector consists of two numbers, the fisrt for min , the second for max}

\item{col.breaks}{break numbers of the color bar, if given, a vector consists of numbers, notes the numbers in the limits}

\item{col.labels}{labs of the break numbers , be consistent with the col.breaks}

\item{size.limits}{limit number of the size bar, default NULL, if given, a vector consists of two numbers, the fisrt for min , the second for max}

\item{size.breaks}{break numbers of the size bar, if given, a vector consists of numbers, notes the numbers in the limits}

\item{size.labels}{labs of the break numbers , be consistent with the size.breaks}
}
\value{
a ggplot2 object of matchstick plot
}
\description{
matchstick_plot for go,kegg,gsea... analysis
}
\examples{
my_matchstick_plot(df = GOdatframe, top = 20, y.decreasing = T,
x = 'Description', y = 'Count', size = 'Count', fill_col = 'pvalue',
col.breaks = c(3,4,5) , col.labels = c('low', '', 'high'),size.limits = c(20,40), size.breaks = c(20,30,40))
}
