% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseq_DEG_analysis.R
\name{my_pca}
\alias{my_pca}
\title{pca visulization for RNAseq grouping}
\usage{
my_pca(dat, group, pal = (ggsci::pal_jco(alpha = 0.7))(9))
}
\arguments{
\item{dat}{raw count matrix (genes are rows, columns are samples)}

\item{group}{group of samples; consistent with samples as column of the matrix}

\item{pal}{color of the groups;number of the colors should be consistent with number of levels of the group}
}
\value{
pca plot
}
\description{
pca visulization for RNAseq grouping
}
\examples{
my_pca(dat, group)
}
