% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_plotting.R
\name{my_themeFULL}
\alias{my_themeFULL}
\title{theme set with 4 axis line around, based on theme_bw}
\usage{
my_themeFULL(
  legend.position = "right",
  baseSize = 8,
  legendSize = 8,
  xSize = 10,
  ySize = 10,
  x.angle = 0,
  x.vjust = 0.5
)
}
\arguments{
\item{legend.position}{legend position}

\item{baseSize}{base text size}

\item{legendSize}{legend size}

\item{xSize}{x lab size}

\item{ySize}{y lab size}

\item{x.angle}{x lab angle}

\item{x.vjust}{x lab vertical position}
}
\value{
a ggplot theme object
}
\description{
theme set with 4 axis line around, based on theme_bw
}
\examples{
my_themeFULL(); my_themeFULL(legend.position = 'top')
}
