% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_plotting.R
\name{my_themeHALF}
\alias{my_themeHALF}
\title{theme set with 2 axis line around ,bottom and left}
\usage{
my_themeHALF(
  legend.position = "bottom",
  titleSize = 15,
  legendSize = 8,
  xlab.Size = 10,
  ylab.Size = 10,
  x.title.Size = 12,
  y.title.Size = 12,
  xlab.angle = 0,
  x.vjust = 0.5
)
}
\arguments{
\item{legend.position}{legend position}

\item{titleSize}{title text size}

\item{legendSize}{legend size}

\item{xlab.Size}{x lab size}

\item{ylab.Size}{y lab size}

\item{x.title.Size}{x axis title size}

\item{y.title.Size}{y axis title size}

\item{xlab.angle}{x lab angle}

\item{x.vjust}{x lab vertical position}
}
\value{
a ggplot theme object
}
\description{
theme set with 2 axis line around ,bottom and left
}
\examples{
my_themeHALF(); my_themeHALF(legend.position = 'top')
}
