% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_plotting.R
\name{my_vocano1}
\alias{my_vocano1}
\title{vocano plot for DEG dataframe}
\usage{
my_vocano1(
  DEG.data,
  x,
  y,
  group,
  title = "",
  x.name = "log2 (Fold Change)",
  y.name = "-log10 (P-value)",
  lab_text = "",
  x_intercept = 1,
  y_intercept = -log10(0.05),
  pal = (ggsci::pal_jco(alpha = 0.7))(9),
  plot_theme
)
}
\arguments{
\item{DEG.data}{DEG dataframe (rownanmes genesymbol; colnames including log2FoldChange(need log2 transformed), pvalue raw value)}

\item{x}{column names meaning log2FoldChange}

\item{y}{column names meaning pvalue (raw value); transform for -log10P in the function}

\item{group}{factors of differentiation, including:'up','stable','down')}

\item{title}{plot title; deflaut blank}

\item{x.name}{plot x lab; deflaut log2 (Fold Change)}

\item{y.name}{plot y lab; deflaut -log10 (P-value)}

\item{lab_text}{Gene names need to be labled}

\item{x_intercept}{vline for lableing mininal log2foldchang3}

\item{y_intercept}{hline for lableing mininal -log10(pvalue)}

\item{pal}{color, default jco}

\item{plot_theme}{theme function returned object (e.g., my_themeFULL(), or theme_bw())}
}
\value{
ggplot2 object
}
\description{
vocano plot for DEG dataframe
}
\examples{
 my_vocano1(DEG.data  = df, x = 'log2FoldChange', y = 'padj', group = 'Dif', x.name = 'log2(Foldchange)', y.name = '-log10(P.value)', lab_text = 'ASNS')
    my_vocano1(DEG.data  = df, x = 'log2FoldChange', y = 'padj', group = 'Dif',pal = c("#DC143C", "#808080","#00008B"), plot_theme = my_themeFULL())
}
